v0.2
I have added an object creator on a grid that auto scales objects as per the grid requirement.
I used Unity's canvas system to implement this. As it can be recoded to achieve this behavior and will also play a part in running the game seamless on all platforms.
Currently the grid manager, tells if a grid is possible for gameplay.