v0.4_Game_Code
In this version, game complete is complete, The grid is working as pr row and columns.
I have now made an algorithm that shuffles key value pairs, irrespective of their length or dimensions. And matches them, winning conditions are also made.
It now is complete game code, the game is starting to look at little alive.